feat(ui): implement Workspace Shape tab (warren-3de4)
Port the live Leveret conversation surface into the Workspace detail
Shape tab. Extract the conversation split-view body (streamed Chat with
the conversation send-message override + dynamic intent editor + send-off)
from ConversationDetail.tsx into a shared ConversationSplitView
(conversation-detail/conversation-surface.tsx) so the /leveret/:id page
and the Shape tab render an identical surface.

ShapeTab resolves the Plot's conversation via conversationsApi.list({plot}),
preferring the active one (else the latest, which goes read-only/quiet
post send-off), exposes Re-wake + the operator-gated Dispatch plan
controls, and renders a Start-conversation affordance when none exists.

Update the structural split-view test to read body assertions from the
shared surface and add a Shape-tab structural test.

pl-0008 step 7
