+++ .mulch/expertise/ui.jsonl
@@ -105,3 +105,4 @@
 {"type":"convention","classification":"tactical","recorded_at":"2026-06-14T02:04:53.565Z","evidence":{"commit":"b74a11ab13cd2d029ab16093d87198c6a4db8b78"},"content":"warren-af60 (pl-0008 step 4): the dead interactive-panel.tsx UI panel + Plots.tsx StartBrainstormDialog were already deleted by the coupled warren-b265 API-client sweep (commit b74a11ab) to keep tsc green ('mutually coupled'). The residual scope for af60 was stale-comment hygiene only: drop dangling references to the deleted interactive-panel.tsx in plot-detail/helpers.ts, plot-detail/run-plan.tsx, and PlotDetail.tsx header comments, plus the formalize-summary non-goal line in components/Chat.tsx and the 'Start brainstorming' affordance note in components/DefaultLanding.tsx. No UI tests referenced the removed panels (only 2 .test files exist under src/ui).","id":"mx-cb93e4"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-06-14T02:11:43.221Z","evidence":{"seeds":"warren-dc54","commit":"678f6284619e5001935c234a33a2f3da10538ec0"},"content":"WorkspacePage (src/ui/src/pages/Workspace.tsx, warren-dc54 / pl-0008 step 5) is the /workspace route: one row per Plot (durable spine) merging the old Plots + Leveret lists. Joins plotsApi.list (server ?status / ?filter=needs_attention chip contract preserved) with conversationsApi.list({}) resolved client-side to the most-recent active conversation per plotId. Columns: name, project, status+needs-you reasons, intent preview, active-conversation indicator (links to /workspace/:id?tab=shape), last activity. Actions: NewConversationButton + NewPlotButton. The shared New Plot dialog lives in src/ui/src/components/NewPlotButton.tsx (NewPlotButton + NewPlotDialog, destination prop defaults /plots) — extracted from Plots.tsx so both /plots and /workspace drive the same create flow.","id":"mx-f5ada8"}
 {"type":"convention","classification":"tactical","recorded_at":"2026-06-14T02:18:33.060Z","evidence":{"commit":"a4a7f0ca94de3d480de77323f6d5c04be290cdfb"},"content":"WorkspaceDetailPage (src/ui/src/pages/WorkspaceDetail.tsx) is the /workspace/:id route — keyed by plotId, loads via plotsApi.get with 5s refetch, reuses PlotDetail's 404/RefreshProjectsCTA recovery branch. Persistent header (PlotNameEditor + StatusTransitionControl + PlotSyncButton + project/summary links) above Shape/Plan/Run/Activity tabs; tab state lives in a ?tab= query param (useSearchParams, default 'shape'). Tab panels are placeholder stubs to be filled by pl-0008 steps 7-10 (warren-3de4/e33f/d17f/ef97). Registered in App.tsx after the /workspace list route.","id":"mx-4beec3"}
+{"type":"pattern","classification":"tactical","recorded_at":"2026-06-14T02:27:28.432Z","evidence":{"commit":"55a0a6a782519d80e4318c73f247d6138c9e1464"},"name":"workspace-shape-tab-reuses-conversation-surface","description":"WorkspaceDetail Shape tab (warren-3de4 / pl-0008 step 7) reuses the shared ConversationSplitView (src/ui/src/pages/conversation-detail/conversation-surface.tsx) — extracted from ConversationDetail.tsx so the /leveret/:id page and the Workspace Shape tab render an identical split-view (streamed Chat with sendMessage->POST /conversations/:id/messages + dynamic fieldsFromIntent/patchFromFields intent editor + SendOffButton). ShapeTab resolves the Plot's conversation via conversationsApi.list({plot:plot.id}), preferring status==='active' then the latest row (post-send-off closed -> surface renders read-only/quiet); NewConversationButton affordance shows when none exist. Structural tests in src/ui-tests now read split-view body assertions from conversation-surface.tsx.","files":["src/ui-tests/conversation-detail-split-view.test.ts","src/ui/src/pages/ConversationDetail.tsx","src/ui/src/pages/WorkspaceDetail.tsx"],"id":"mx-d86a04"}
